requirements.txt
fastapi
uvicorn
numpy
Pillow
scikit-learn
transformers
torch
torchvision
certifi
